Further optimization of DBID layer and heaps. Fixes a performance regressions (and speeds up some more) due to odd optimization problem of hotspot JIT. Enhance heap by adding a "replaceTopElement" option [closes ticket:96] This is not the best solution, but it is the fastest (and comparably reasonable) after days of benchmarking.

+++ src/de/lmu/ifi/dbs/elki/database/datastore/memory/ArrayIntegerStore.java
@@ -77,7 +77,7 @@ public class ArrayIntegerStore implements WritableIntegerDataStore {
   @Deprecated
   public Integer get(DBIDRef id) {
     try {
-      return data[idmap.mapDBIDToOffset(id)];
+      return Integer.valueOf(data[idmap.mapDBIDToOffset(id)]);
     }
     catch(ArrayIndexOutOfBoundsException e) {
       return null;
@@ -89,8 +89,8 @@ public class ArrayIntegerStore implements WritableIntegerDataStore {
   public Integer put(DBIDRef id, Integer value) {
     final int off = idmap.mapDBIDToOffset(id);
     int ret = data[off];
-    data[off] = value;
-    return ret;
+    data[off] = value.intValue();
+    return Integer.valueOf(ret);
   }
   
   @Override
